+++ apps/web/app/(authenticated)/mentorship/logs/create/page.tsx
@@ -1,0 +1,13 @@
+import NewInteractionForm from "@/app/lib/components/InteractionLogs/new-interaction-form";
+
+
+export default async function Create() {
+
+
+    return (
+        <div>
+            <NewInteractionForm />
+        </div>
+
+    )
+}
