+++ apps/web/app/lib/components/InteractionLogs/edit-button.tsx
@@ -1,25 +1,10 @@
-"use client";
-
 import { Pencil } from 'lucide-react';
-import { useRouter, usePathname, useSearchParams } from 'next/navigation';
+import Link from 'next/link';
 
-export default function EditButton({ selectedId }: { selectedId?: string | string[] | undefined }) {
-    const router = useRouter();
-    const pathname = usePathname();
-    const searchParams = useSearchParams();
-
-    const handleEdit = () => {
-        const params = new URLSearchParams(searchParams.toString());
-        if (Array.isArray(selectedId)) {
-            selectedId = selectedId[0]
-        }
-        if (selectedId) {
-            params.set('edit', selectedId);
-        }
-        router.replace(`${pathname}?${params.toString()}`);
-    };
-
+export default function EditButton({ selectedId }: { selectedId?: string | undefined }) {
     return (
-        <Pencil onClick={handleEdit} />
+        <Link href={`/mentorship/logs/edit?id=${selectedId}`}>
+            <Pencil />
+        </Link>
     );
 }
